rustic::app::prelude

Struct FSConfig

pub struct FSConfig {
    pub root_dir: PathBuf,
    pub score_path: PathBuf,
    pub instrument_path: PathBuf,
    pub recordings_path: PathBuf,
}

Application Filesystem configuration

Fields
root_dir: PathBufscore_path: PathBufinstrument_path: PathBufrecordings_path: PathBuf
Implementations

impl FSConfig

pub fn debug_dir(module: &str, filename: &str) -> Result<PathBuf, ()>

Builds the required path to save the file from the given module with the given name

Returns a result containing the built path with the file name.

pub fn _stamped_debug_dir(module: &str, filename: &str) -> Result<PathBuf, ()>

Builds the required path to save the file from the given module with the given name. Adds a timestamp to allow for time-differentiation of the saved files.

Returns a result containing the built path with the file name.

pub fn app_root_dir() -> Result<PathBuf, String>

Returns the app’s default root path for saving configuration files & other. This is supposed to be used if the application’s settings structure contains no information about the path.