rustic::audio::render_thread

Function spawn_audio_render_thread

pub fn spawn_audio_render_thread(
    shared_state: Arc<SharedAudioState>,
    instruments: Vec<Box<dyn Instrument + Send + Sync>>,
    message_rx: Receiver<AudioMessage>,
    audio_queue: Arc<ArrayQueue<f32>>,
    config: AudioConfig,
) -> JoinHandle<()>

Spawns the audio render thread that generates audio samples

This thread:

Processes control messages from the command thread

Generates audio by calling instrument methods

Writes audio to the ring buffer for the cpal callback to consume