rustic::core::envelope

Module prelude

Structs

ADSREnvelope

ADSREnvelopeBuilder

BezierSegment

Represents a Bezier Segment for an envelope. A bezier segment is defined by a starting point, an ending point and a control point. Since segments times are normalized between 0.0 and 1.0, the x axis values for the start and end points are, respectively, 0.0 and 1.0.

ConstantSegment

A constant segment for the sustain phase of an envelope.

FunctionSegment

A segment of an envelope defined by a custom function. Note: the function field is not serializable. When deserialized, it defaults to the identity function |t| t. This segment type exists primarily for code-level use, not TOML definitions.

LinearSegment

A simple linear segment for an envelope. The segment interpolates linearly between the start and end values over the specified duration.

Traits

Segment

A segment is a part of an envelope that maps a time range to amplitude values. The time range is normalized between 0.0 and 1.0 for the duration of the segment.

SustainSegment

A sustain segment is a special kind of segment that does not define a time range, but can be queried for as long as the note is held. It can be a constant value or any other function in R+.