rustic::core::envelope::prelude

Trait Segment

pub trait Segment:
    Display
    + DynClone
    + Debug
    + Send
    + Sync
    + Serialize
    + Deserialize {
    // Required methods
    fn at(&self, time: f32) -> f32;
    fn get_duration(&self) -> f32;

    // Provided method
    fn map_time(&self, segment_start: f32, current_time: f32) -> f32 { ... }
}

A segment is a part of an envelope that maps a time range to amplitude values. The time range is normalized between 0.0 and 1.0 for the duration of the segment.

It is up to the envelope implementation to manage the time mapping and sequencing of segments. For convenience, the segment trait provides a method to map the global time to the segment’s local time.

Required Methods

fn at(&self, time: f32) -> f32

Returns the amplitude at the given time position within the segment. Any value of time above one will be clamped to the end value of the segment. (time > 1.0 ? 1.0 : time)

fn get_duration(&self) -> f32

Returns the duration in seconds of the segment.

Provided Methods

fn map_time(&self, segment_start: f32, current_time: f32) -> f32

Maps the global time to the segment’s local time [0.0 - 1.0] Any returned value above 1.0 means that the segment has completed.