rustic::core::envelope

Trait Envelope

pub trait Envelope:
    Display
    + Debug
    + Send
    + Sync
    + Serialize
    + Deserialize {
    // Required methods
    fn at(&self, time: f32, note_off: f32) -> f32;
    fn completed(&self, time: f32, note_off: f32) -> bool;
}

An envelope that can be used to modulate a signal over time. The base principle is simply to have a function with a varying value over time. This value can then be used to shape either the amplitude, frequency or any other parameter of a sound.

Required Methods

fn at(&self, time: f32, note_off: f32) -> f32

Returns the envelope value at the given point in time. The timestamps is expected to be mapped to the envelope’s duration, that is the minimum value is 0.0.

fn completed(&self, time: f32, note_off: f32) -> bool

Returns whether the envelope has completed or not based on the current time & note_off timestamp