rustic::core::generator::prelude

Enum MixMode

pub enum MixMode {
    Sum,
    Multiply,
    Max,
    Average,
}

A mixing mode for combining multiple tone generators in the CompositeGenerator.

Sum: Adds the outputs of all tone generators together.

Multiply: Multiplies the outputs of all tone generators together.

Max: Takes the maximum output value from all tone generators.

Average: Averages the outputs of all tone generators.

Variants

Sum

Multiply

Max

Average