rustic::core

Module graph

Audio signal graph processing and routing Graphs are used to build audio pipelines for the application. They serve as metastructures for filters. Graphs can be used to create complex audio effects and processing chains.

Structs

SimpleSink

A simple audio sink that stores incoming audio samples. (Allowing other parts of the code to pull its values)

SimpleSource

System

The system module contains the implementation of the system element. A Pipe & Filter system The system is composed of filters, sources and sinks. It is represented as a directed graph where the filters are the nodes The Sources & Sinks are special nodes that have respectively only outgoing (source) or incoming (sink) edges The edges represent the pipes between the filters Some filters have special output properties. E.g. the delay filter’s input pipe is ignored (postponed) when the topology sorting is done, in order to avoid cycles. A system with cycles must include a delay or similar filter to break the cycle.

Traits

AudioGraphElement

An element in the Audio pipeline Has a name and an index. Is able to connect to other elements

Entry

A trait for AudioGraphElements to allow other elements to push values into them.

Filter

A filter that can process data. Data should be pushed to the filter’s input by either the preceding filter or a source.

Sink

A trait for AudioGraphElements that allow other parts of the code to consume values from them. (Acts as a graph output)

Source

A trait for AudioGraphElements that allow neighbours in the graph to pull values from them. (Acts as a graph input)

Functions

simple_source

Creates a simple source with a given generator and frequency