rustic::core::graph

Trait Entry

pub trait Entry:
    Debug
    + DynClone
    + Send {
    // Required method
    fn push(&mut self, block: Arc<Block>, port: usize);
}

A trait to allow other elements to push values into them.

Required Methods

fn push(&mut self, block: Arc<Block>, port: usize)

Pushes a block into this element on port port