rustic::core::graph

Trait Filter

pub trait Filter:
    Entry
    + Display
    + MetaFilter
    + Send
    + Sync {
    // Required methods
    fn transform(&mut self) -> Vec<Block> ;
    fn as_any_mut(&mut self) -> &mut dyn Any;

    // Provided method
    fn postponable(&self) -> bool { ... }
}

A filter that can process data. Data should be pushed to the filter’s input by either the preceding filter or a source.

Required Methods

fn transform(&mut self) -> Vec<Block>

Applies the filter’s transformation to the input Returns a Vector of Blocks. Each block correspond to output port x of the filter

fn as_any_mut(&mut self) -> &mut dyn Any

Enables downcasting from trait object to concrete type.

Provided Methods

fn postponable(&self) -> bool

Returns true if the filter’s execution can be postponed to the end of the execution cycle of the graph. A postponable element must be present in a cycle of the graph to avoid infinite looping. E.g. a delay filter can be postponed if it lies within a feedback loop.