rustic::score::score

Struct Score

pub struct Score {
    pub name: String,
    pub signature: TimeSignature,
    pub tempo: usize,
    pub staves: Vec<Staff>,
    pub instruments: Vec<Box<dyn Instrument>>,
}

A music score. Has a defined time signature, tempo, staves associated with their instruments, name, …

Fields
name: Stringsignature: TimeSignaturetempo: usizestaves: Vec<Staff>instruments: Vec<Box<dyn Instrument>>
Implementations

impl Score

pub fn new<S: AsRef<str>>( name: S, signature: TimeSignature, tempo: usize, instruments: Vec<Box<dyn Instrument>>, staves: Vec<Staff>, ) -> Self

pub fn load_toml(path: &Path) -> Result<Self, String>

Loads a score from a toml file. Returns either a score or an error message.

pub fn dump_toml(&self) -> Result<String, String>

Dumps the score to a toml string.

pub fn save(&self, path: &Path) -> Result<(), String>

Saves the score to a toml file.

pub fn add_staff(&mut self) -> usize

Adds a new staff to the score and returns its index. The staff will not be associated with any instrument.

pub fn add_instrument(&mut self, instrument: Box<dyn Instrument>) -> usize

Adds an instrument to the score and returns its index. The instrument’s index is also added to the corresponding staff. If the number of staves is less than the number of instruments, a new staff is created.

use rustic::prelude::{Score, TimeSignature};
use rustic::instruments::prelude::{HiHat, Kick, Snare};

let mut score = Score::new("Test", TimeSignature(4, 4), 120, Vec::new(), Vec::new());
let kick_index = score.add_instrument(Box::new(Kick::new()));

pub fn add_note(&mut self, instrument: usize, note: Note) -> Result<(), String>

Adds a note to the score for the specified instrument at the first available space.

pub fn play(&mut self) -> Result<(), String>

Plays the score by compiling it first and then playing it.

This method creates an optimized CompiledScore representation of the score, plays it, and then returns the instruments back to the score.

Returns

Ok(()) - If playback completed successfully

Err(String) - If an error occurred during playback

Example

use rustic::prelude::{Score, TimeSignature};
use rustic::instruments::prelude::HiHat;

let mut score = Score::new("Demo", TimeSignature(4, 4), 120, Vec::new(), Vec::new());
score.add_instrument(Box::new(HiHat::new().unwrap()));

// Play the score
score.play().unwrap();