rustic::core::graph

Struct System

pub struct System<const INPUTS: usize, const OUTPUTS: usize> { /* private fields */ }

The system module contains the implementation of the system element. A Pipe & Filter system The system is composed of filters, sources and sinks. It is represented as a directed graph where the filters are the nodes The Sources & Sinks are special nodes that have respectively only outgoing (source) or incoming (sink) edges The edges represent the pipes between the filters Some filters have special output properties. E.g. the delay filter’s input pipe is ignored (postponed) when the topology sorting is done, in order to avoid cycles. A system with cycles must include a delay or similar filter to break the cycle.

use rustic::core::graph::System;
use rustic::core::filters::prelude::Tremolo;

// A simple system with one input and one output
let mut system = System::<1, 1>::new();

// Adding a filter to the system
let filter = Tremolo::new(20.0, 0.5, 1.5);
let filter_index = system.add_filter(Box::from(filter));

Implementations

impl<const INPUTS: usize, const OUTPUTS: usize> System<INPUTS, OUTPUTS>

pub fn new() -> Self

Creates a new system with simple null sources & simple sinks

pub fn merge<const T: usize>( self, other: System<OUTPUTS, T>, mapping: Vec<(usize, usize)>, ) -> System<INPUTS, T>

Merges the two systems together to create a new one. The graphs are merged following the given mapping from sinks to sources. Sinks to sources links are replaced with a simple combinator filter. The amount of input in the second system should match the amount of output in the first system.

pub fn add_filter(&mut self, filter: Box<dyn Filter>) -> NodeIndex<u32>

pub fn connect( &mut self, from: NodeIndex<u32>, to: NodeIndex<u32>, out_port: usize, in_port: usize, )

pub fn connect_source( &mut self, source: usize, to: NodeIndex<u32>, in_port: usize, )

Connects a source to a filter of the graph

pub fn connect_sink( &mut self, from: NodeIndex<u32>, sink: usize, out_port: usize, )

Connects a filter from the graph to a sink

pub fn set_sink(&mut self, index: usize, sink: Box<dyn Sink>)

Sets the sink at index index to be the given sink object

pub fn set_source(&mut self, index: usize, source: Box<dyn Source>)

Sets source at index index to be the given source object

pub fn compute(&mut self) -> Result<(), ()>

pub fn run(&mut self)

pub fn get_sink(&mut self, index: usize) -> Result<&mut Box<dyn Sink>, &str>

Returns a sink pipe from the system. If the index is out of bounds, returns an error.

pub fn get_filter_mut( &mut self, index: NodeIndex<u32>, ) -> Option<&mut Box<dyn Filter>>

Returns a mutable reference to a filter in the graph by its NodeIndex. This allows direct access to filter-specific methods (like reset()) that aren’t part of the Filter trait.

pub fn save_to_file(&self, path: &Path) -> Result<(), String>