rustic::core::graph

Trait Filter

pub trait Filter:
    Entry
    + AudioGraphElement
    + Display
    + Debug {
    // Required methods
    fn transform(&mut self) -> Vec<f32>;
    fn postponable(&self) -> bool;
    fn as_any_mut(&mut self) -> &mut dyn Any;
}

A filter that can process data. Data should be pushed to the filter’s input by either the preceding filter or a source.

Required Methods

fn transform(&mut self) -> Vec<f32>

Applies the filter’s transformation to the input Returns a tuple of the output and the indices of the elements that the filter is connected to

fn postponable(&self) -> bool

Returns true if the filter’s execution can be postponed to the end of the execution cycle of the graph. An postonable element must be present in a cycle of the graph to avoid infinite looping. E.g. a delay filter can be postponed if it lies within a feedback loop.

fn as_any_mut(&mut self) -> &mut dyn Any

Enables downcasting from trait object to concrete type. This is required to access type-specific methods (like reset()) that aren’t part of the Filter trait.